adding photo detail in n/w layer.
