+++ app/src/main/java/com/karthik/splash/photodetailscreen/PhotoDetailScreenViewModel.kt
@@ -2,10 +2,8 @@ package com.karthik.splash.photodetailscreen
 
 import androidx.lifecycle.*
 import com.karthik.network.IMemoryCache
-import com.karthik.splash.models.likephoto.LikeResponse
-import com.karthik.splash.models.photodetail.PhotoDetailInfo
-import com.karthik.splash.photodetailscreen.network.PhotoDetailsResponse
-import com.karthik.splash.photodetailscreen.network.PhotoLikeResponse
+import com.karthik.network.photodetailscreen.IPhotoDetailScreenRepository
+import com.karthik.network.photodetailscreen.models.*
 import kotlinx.coroutines.launch
 
 @Suppress("UNCHECKED_CAST")
@@ -24,16 +22,20 @@ class PhotoDetailScreenViewModel(
 ) : ViewModel() {
     private val details: MutableLiveData<PhotoDetailInfo> = MutableLiveData()
     private val like: MutableLiveData<LikeResponse> = MutableLiveData()
+    private val internalState: MutableLiveData<PhotoDetailsNetworkState> = MutableLiveData()
 
     val photolike: LiveData<LikeResponse> = like
     val photodetails: LiveData<PhotoDetailInfo> = details
+    val networkState: LiveData<PhotoDetailsNetworkState> = internalState
 
     fun getPhotoDetail(id: String) {
         viewModelScope.launch {
             val photoInfoResponse = photoRepository.getPhotoInfo(id)
             if (photoInfoResponse is PhotoDetailsResponse.PhotoDetailsSuccessResponse) {
+                internalState.postValue(PhotoDetailsNetworkState.PhotoDetailsNetworkLoadSuccess)
                 details.postValue(photoInfoResponse.photoDetail)
             } else {
+                internalState.postValue(PhotoDetailsNetworkState.PhotoDetailsNetworkLoadError)
                 //TODO:handle error
             }
         }
@@ -43,16 +45,14 @@ class PhotoDetailScreenViewModel(
         viewModelScope.launch {
             val photoLikeResponse = photoRepository.likePhoto(id)
             if (photoLikeResponse is PhotoLikeResponse.PhotoDetailsSuccessResponse) {
+                internalState.postValue(PhotoDetailsNetworkState.PhotoDetailsNetworkLoadSuccess)
                 like.postValue(photoLikeResponse.likeResponse)
             }else{
+                internalState.postValue(PhotoDetailsNetworkState.PhotoLikeNetworkLoadError)
                 //TODO:handle error
             }
         }
     }
 
-    fun isUserLoggedIn() =
-            memoryCache.isUserLoggedIn()
-
-    fun getnetworkState() =
-            photoRepository.getPhotoDetailNetworkState()
+    fun isUserLoggedIn() = memoryCache.isUserLoggedIn()
 }
