+++ app/src/main/java/com/karthik/splash/photodetailscreen/di/PhotoDetailScreenModule.kt
@@ -1,10 +1,10 @@
 package com.karthik.splash.photodetailscreen.di
 
 
-import com.karthik.splash.photodetailscreen.IPhotoDetailScreenRepository
-import com.karthik.splash.photodetailscreen.network.PhotoDetailScreenRepository
+
+import com.karthik.network.photodetailscreen.IPhotoDetailScreenRepository
+import com.karthik.network.photodetailscreen.repository.PhotoDetailScreenRepository
 import com.karthik.splash.photodetailscreen.PhotoDetailScreenViewModelFactory
-import com.karthik.splash.photodetailscreen.network.PhotoService
 import com.karthik.splash.storage.MemoryCache
 import dagger.Module
 import dagger.Provides
@@ -15,8 +15,7 @@ class PhotoDetailScreenModule {
 
     @Provides
     fun providesPhotoNetworkLayer(retrofit: Retrofit): IPhotoDetailScreenRepository {
-        val photoService: PhotoService = retrofit.create(PhotoService::class.java)
-        return PhotoDetailScreenRepository(photoService = photoService)
+        return PhotoDetailScreenRepository(retrofit)
     }
 
     @Provides
