+++ network/src/main/java/com/karthik/network/photodetailscreen/IPhotoDetailScreenRepository.kt
@@ -1,0 +1,14 @@
+package com.karthik.network.photodetailscreen
+
+import com.karthik.network.photodetailscreen.models.PhotoDetailsResponse
+import com.karthik.network.photodetailscreen.models.PhotoLikeResponse
+
+interface IPhotoDetailScreenRepository {
+    suspend fun getPhotoInfo(
+            id: String
+    ): PhotoDetailsResponse
+
+    suspend fun likePhoto(
+            id: String
+    ): PhotoLikeResponse
+}
