+++ network/src/main/java/com/karthik/network/photodetailscreen/repository/PhotoDetailsNetworkService.kt
@@ -1,7 +1,7 @@
-package com.karthik.splash.photodetailscreen.network
+package com.karthik.network.photodetailscreen.repository
 
-import com.karthik.splash.models.likephoto.LikeResponse
-import com.karthik.splash.models.photodetail.PhotoDetailInfo
+import com.karthik.network.photodetailscreen.models.LikeResponse
+import com.karthik.network.photodetailscreen.models.PhotoDetailInfo
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.POST
@@ -11,7 +11,7 @@ import retrofit2.http.Path
  * Created by karthikrk on 09/12/17.
  */
 
-interface PhotoService {
+internal interface PhotoDetailsNetworkService {
     @GET("photos/{id}")
     suspend fun getPhotoInfo(@Path("id") id: String): Response<PhotoDetailInfo>
